+++ R/detect.R
@@ -108,39 +108,77 @@ str_ends <- function(string, pattern, negate = FALSE) {
   )
 }
 
-#' Detect a pattern in the same way as `SQL`'s `LIKE` operator
+#' Detect a pattern in the same way as `SQL`'s `LIKE` and `ILIKE` operators
 #'
 #' @description
-#' `str_like()` follows the conventions of the SQL `LIKE` operator:
+#' `str_like()` and `str_like()` follow the conventions of the SQL `LIKE`
+#' and `ILIKE` operators, namely:
 #'
 #' * Must match the entire string.
 #' * `_` matches a single character (like `.`).
 #' * `%` matches any number of characters (like `.*`).
 #' * `\%` and `\_` match literal `%` and `_`.
-#' * The match is case insensitive by default.
+#'
+#' The difference between the two functions is their case-sensitivity:
+#' `str_like()` is case sensitive and `str_ilike()` is not.
+#'
+#' @note
+#' Prior to stringr 1.6.0, `str_like()` was incorrectly case-insensitive.
 #'
 #' @inheritParams str_detect
 #' @param pattern A character vector containing a SQL "like" pattern.
 #'   See above for details.
-#' @param ignore_case Ignore case of matches? Defaults to `TRUE` to match
-#'   the SQL `LIKE` operator.
+#' @param ignore_case `r lifecycle::badge("deprecated")`
 #' @return A logical vector the same length as `string`.
 #' @export
 #' @examples
 #' fruit <- c("apple", "banana", "pear", "pineapple")
 #' str_like(fruit, "app")
 #' str_like(fruit, "app%")
+#' str_like(fruit, "APP%")
 #' str_like(fruit, "ba_ana")
-#' str_like(fruit, "%APPLE")
-str_like <- function(string, pattern, ignore_case = TRUE) {
+#' str_like(fruit, "%apple")
+#'
+#' str_ilike(fruit, "app")
+#' str_ilike(fruit, "app%")
+#' str_ilike(fruit, "APP%")
+#' str_ilike(fruit, "ba_ana")
+#' str_ilike(fruit, "%apple")
+str_like <- function(string, pattern, ignore_case = deprecated()) {
+  check_lengths(string, pattern)
+  check_character(pattern)
+  if (inherits(pattern, "stringr_pattern")) {
+    cli::cli_abort("{.arg pattern} must be a plain string, not a stringr modifier.")
+  }
+  if (lifecycle::is_present(ignore_case)) {
+    lifecycle::deprecate_warn(
+      when = "1.6.0",
+      what = "str_like(ignore_case)",
+      details = c(
+        "`str_like()` is always case sensitive.",
+        "Use `str_ilike()` for case insensitive string matching."
+      )
+    )
+    check_bool(ignore_case)
+    if (ignore_case) {
+      return(str_ilike(string, pattern))
+    }
+  }
+
+  pattern <- regex(like_to_regex(pattern), ignore_case = FALSE)
+  stri_detect_regex(string, pattern, opts_regex = opts(pattern))
+}
+
+#' @export
+#' @rdname str_like
+str_ilike <- function(string, pattern) {
   check_lengths(string, pattern)
   check_character(pattern)
   if (inherits(pattern, "stringr_pattern")) {
     cli::cli_abort(tr_("{.arg pattern} must be a plain string, not a stringr modifier."))
   }
-  check_bool(ignore_case)
 
-  pattern <- regex(like_to_regex(pattern), ignore_case = ignore_case)
+  pattern <- regex(like_to_regex(pattern), ignore_case = TRUE)
   stri_detect_regex(string, pattern, opts_regex = opts(pattern))
 }