+++ tests/testthat/test-detect.R
@@ -55,9 +55,24 @@ test_that("functions use tidyverse recycling rules", {
 # str_like ----------------------------------------------------------------
 
 
-test_that("str_like works", {
+test_that("str_like is case sensitive", {
   expect_true(str_like("abc", "ab%"))
+  expect_false(str_like("abc", "AB%"))
   expect_snapshot(str_like("abc", regex("x")), error = TRUE)
+})
+
+test_that("ignore_case is deprecated but still respected", {
+  expect_snapshot(out <- str_like("abc", "AB%", ignore_case = TRUE))
+  expect_equal(out, TRUE)
+
+  expect_warning(out <- str_like("abc", "AB%", ignore_case = FALSE))
+  expect_equal(out, FALSE)
+})
+
+test_that("str_ilike works", {
+  expect_true(str_ilike("abc", "ab%"))
+  expect_true(str_ilike("abc", "AB%"))
+  expect_snapshot(str_ilike("abc", regex("x")), error = TRUE)
 })
 
 test_that("like_to_regex generates expected regexps",{