+++ NEWS.md
@@ -1,5 +1,7 @@
 # stringr (development version)
 
+* New `str_ilike()` that follows the conventions of the SQL ILIKE operator (@edward-burn, #543).
+* `str_like(ignore_case)` is deprecated, with `str_like()` now always case sensitive to better follow the conventions of the SQL LIKE operator (@edward-burn, #543).
 * `str_sub<-` now gives a more informative error if `value` is not the correct length.
 * Add `sep` argument to `str_dup()` so that it is possible to repeat a string and
   add a separator between every repeated value (@edward-burn, #564).